requirements.txt
alpaca-trade-api
pandas
numpy
torch
gym
python-dotenv
yfinance
pytz